Move muting function over to video store
Makes more sense here!
