+++ src/MuteButton.js
@@ -1,23 +1,23 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { toggleMute } from './actions/sessionActions';
+import { toggleMute } from './actions/videoActions';
 
 class MuteButton extends Component {
   toggleMute = () => {
-    this.props.toggleMute(this.props.session.muted);
+    this.props.toggleMute(this.props.video.muted);
   }
 
   render() {
     return (
       <button onClick={this.toggleMute}>
-        {this.props.session.muted ? 'Unmute' : 'Mute'}
+        {this.props.video.muted ? 'Unmute' : 'Mute'}
       </button>
     );
   }
 }
 
 const mapStateToProps = state => ({
-  session: state.session
+  video: state.video
 });
 
 export default connect(mapStateToProps, { toggleMute })(MuteButton);
