+++ src/actions/sessionTypes.js
@@ -1,2 +1,1 @@
 export const SET_SECONDS_UNTIL_NEXT_PROGRAM = 'SET_SECONDS_UNTIL_NEXT_PROGRAM';
-export const TOGGLE_MUTE = 'TOGGLE_MUTE';
