+++ src/actions/videoTypes.js
@@ -1,1 +1,2 @@
 export const ADD_VIDEO_PLAYER = 'ADD_VIDEO_PLAYER';
+export const TOGGLE_MUTE = 'TOGGLE_MUTE';
