+++ src/reducers/sessionReducer.js
@@ -1,10 +1,8 @@
-import { SET_SECONDS_UNTIL_NEXT_PROGRAM, TOGGLE_MUTE } from '../actions/sessionTypes';
+import { SET_SECONDS_UNTIL_NEXT_PROGRAM } from '../actions/sessionTypes';
 
 const initialState = {
   currentHour: new Date().getHours(),
-  secondsUntilNextProgram: null,
-  muted: true,
-  volume: 0
+  secondsUntilNextProgram: null
 }
 
 export default function(state = initialState, action) {
@@ -13,12 +11,6 @@ export default function(state = initialState, action) {
       return {
         ...state,
         secondsUntilNextProgram: action.secondsUntilNextProgram
-      }
-    case TOGGLE_MUTE:
-      return {
-        ...state,
-        muted: action.muted,
-        volume: action.volume
       }
     default:
       return state;
