+++ src/reducers/videoReducer.js
@@ -1,8 +1,8 @@
-import { ADD_VIDEO_PLAYER } from '../actions/videoTypes';
+import { ADD_VIDEO_PLAYER, TOGGLE_MUTE } from '../actions/videoTypes';
 
 const initialState = {
-  // muted: true,
-  // volume: 0,
+  muted: true,
+  volume: 0,
   player: null
 }
 
@@ -12,6 +12,12 @@ export default function(state = initialState, action) {
       return {
         ...state,
         player: action.player
+      }
+    case TOGGLE_MUTE:
+      return {
+        ...state,
+        muted: action.muted,
+        volume: action.volume
       }
     default:
       return state;
